refactor(tunnel): split tunnel cli and make it a standalone package
