+++ packages/tunnel/bin/index.js
@@ -1,0 +1,3 @@
+#!/usr/bin/env node
+// eslint-disable-next-line import/no-unassigned-import
+import '../lib/index.js';
