+++ packages/tunnel/package.json
@@ -1,7 +1,7 @@
 {
-  "name": "@logto/cli",
-  "version": "1.19.0",
-  "description": "Logto CLI.",
+  "name": "@logto/tunnel",
+  "version": "0.0.0",
+  "description": "A CLI tool that creates tunnel service to Logto Cloud for local development.",
   "author": "Silverhand Inc. <contact@silverhand.io>",
   "homepage": "https://github.com/logto-io/logto#readme",
   "license": "MPL-2.0",
@@ -11,7 +11,7 @@
   },
   "main": "lib/index.js",
   "bin": {
-    "logto": "bin/logto.js"
+    "logto-tunnel": "bin/index.js"
   },
   "files": [
     "bin",
@@ -25,7 +25,6 @@
     "precommit": "lint-staged",
     "prepare:package-json": "node -p \"'export const packageJson = ' + JSON.stringify(require('./package.json'), undefined, 2) + ';'\" > src/package-json.ts",
     "build": "rm -rf lib && pnpm prepare:package-json && tsc -p tsconfig.build.json",
-    "build:test": "pnpm build",
     "dev": "tsc -p tsconfig.build.json --watch --preserveWatchOutput --incremental",
     "start": "node .",
     "start:dev": "pnpm build && node .",
@@ -42,43 +41,25 @@
     "url": "https://github.com/logto-io/logto/issues"
   },
   "dependencies": {
-    "@logto/connector-kit": "workspace:^4.0.0",
-    "@logto/core-kit": "workspace:^2.5.0",
-    "@logto/schemas": "workspace:1.19.0",
-    "@logto/shared": "workspace:^3.1.1",
+    "@logto/core-kit": "workspace:^",
+    "@logto/shared": "workspace:^",
     "@silverhand/essentials": "^2.9.1",
-    "@silverhand/slonik": "31.0.0-beta.2",
     "chalk": "^5.3.0",
-    "decamelize": "^6.0.0",
     "dotenv": "^16.4.5",
-    "got": "^14.0.0",
-    "hpagent": "^1.2.0",
-    "inquirer": "^9.0.0",
-    "nanoid": "^5.0.1",
-    "ora": "^8.0.1",
-    "p-limit": "^6.0.0",
-    "p-retry": "^6.0.0",
-    "pg-protocol": "^1.6.0",
-    "semver": "^7.3.8",
-    "tar": "^7.0.0",
+    "http-proxy-middleware": "^3.0.0",
+    "mime": "^4.0.4",
     "yargs": "^17.6.0",
     "zod": "^3.23.8"
   },
   "devDependencies": {
     "@silverhand/eslint-config": "6.0.1",
     "@silverhand/ts-config": "6.0.0",
-    "@types/inquirer": "^9.0.0",
     "@types/node": "^20.9.5",
-    "@types/semver": "^7.3.12",
-    "@types/sinon": "^17.0.0",
-    "@types/tar": "^6.1.12",
     "@types/yargs": "^17.0.13",
     "@vitest/coverage-v8": "^2.0.0",
-    "@withtyped/server": "^0.13.6",
     "eslint": "^8.56.0",
     "lint-staged": "^15.0.0",
     "prettier": "^3.0.0",
-    "sinon": "^18.0.0",
     "typescript": "^5.5.3",
     "vitest": "^2.0.0"
   },
