+++ packages/tunnel/src/index.ts
@@ -3,11 +3,9 @@ import dotenv from 'dotenv';
 import yargs from 'yargs';
 import { hideBin } from 'yargs/helpers';
 
-import connector from './commands/connector/index.js';
-import database from './commands/database/index.js';
-import install from './commands/install/index.js';
+import tunnel from './commands/index.js';
 import { packageJson } from './package-json.js';
-import { cliConfig, ConfigKey, consoleLog } from './utils.js';
+import { consoleLog } from './utils.js';
 
 void yargs(hideBin(process.argv))
   .version(false)
@@ -16,16 +14,13 @@ void yargs(hideBin(process.argv))
     describe: 'The path to your `.env` file',
     type: 'string',
   })
-  .option('db', {
-    alias: ['db-url', 'database-url'],
-    describe: 'The Postgres URL to Logto database',
-    type: 'string',
-  })
   .option('version', {
     alias: 'v',
-    describe: 'Print Logto CLI version',
+    describe: 'Print CLI version',
     type: 'boolean',
-    global: false,
+  })
+  .middleware(({ env }) => {
+    dotenv.config({ path: env });
   })
   .middleware(({ version }) => {
     if (version) {
@@ -34,19 +29,7 @@ void yargs(hideBin(process.argv))
       process.exit(0);
     }
   }, true)
-  .middleware(({ env, db: databaseUrl }) => {
-    dotenv.config({ path: env });
-
-    const initialDatabaseUrl = databaseUrl ?? process.env[ConfigKey.DatabaseUrl];
-
-    if (initialDatabaseUrl) {
-      cliConfig.set(ConfigKey.DatabaseUrl, initialDatabaseUrl);
-    }
-  })
-  .command(install)
-  .command(database)
-  .command(connector)
-  .demandCommand(1)
+  .command(tunnel)
   .showHelpOnFail(false, `Specify ${chalk.green('--help')} for available options`)
   .strict()
   .parserConfiguration({
