+++ packages/tunnel/tsconfig.json
@@ -1,0 +1,10 @@
+{
+  "extends": "@silverhand/ts-config/tsconfig.base",
+  "compilerOptions": {
+    "outDir": "lib",
+    "types": ["node"]
+  },
+  "include": [
+    "src"
+  ]
+}
